fix: add podspec so that pod install doesnt complain
A root podspec file should be published with package so that pod install doesn't complain when installing projects with bob.

I add it here but not quite sure where it should be in order to be published at the root of package.

Also added homepage property in package.json as required by podspec.