`blacklist` -> `exclusionList` migration for examples metro.config.js (#168)
* Update react-native version to 0.64.0

The react-native version needs to be updated to 0.64.x so it will
support the blacklist -> exclusionList migration.
0.64.0 has been chosen as the initial stable version.

NOTE: For an incoming Windows support the 0.64.0 is also the required
version of react-native

* Change blacklist to exclusionList in example's metro config

* Change resolver field to blacklistRE

It turns out that in the metro there's still a blacklist used for a
resolver field. This makes the metro to be unable to find the new word.
To fix that, the blacklistRE field name is brought back.