refactor: use GitHub actions instead of CircleCI for generated project
This adds config for GitHub actions and removes the CircleCI config from the
generated project.

Previously, we kept CircleCI config as it provided better caching and was more
optimized due to multiple jobs being able to reuse the result from the one job,
reducing the setup time.

While the problems with setup time still remains with GitHub actions, the caching
part is better now which allows us to skip installing dependencies - reducing
those concerns. The biggest advantage of GitHub actions is that since it's
integrated with GitHub, the CI doesn't require additional steps after creating a
project.

In addition, we've moved to GitHub actions for this monorepo already, so it's
easier for us to maintain the config for GitHub actions as opposed a CircleCI
config that we don't use.