chore: upgrade babel-plugin-module-resolver to 5.0.0 (#381)
### Summary

The package `babel-plugin-module-resolver` is currently using version
`4.1.0` of the package. One of the dependencies of this package,
`json5`, has security vulnerabilities in the version used in this
version. These vulnerabilities have been fixed in the latest version of
the `babel-plugin-module-resolver` package. So upgrade the package to
the latest version.

See the CHANGELOG of the package for more information.

https://github.com/tleunen/babel-plugin-module-resolver/blob/master/CHANGELOG.md#500-2023-01-09