feat: bump node version to 18 (#418)
### Summary

This bumps the node version of the project to `v18.12.0`(18 LtS)

#### Why?

1. Node.js 16 LTS is coming to end-of-life later this year.
[Source](https://nodejs.org/en/blog/announcements/nodejs16-eol)
2. You are not able to install the dependencies of react native example
app anymore. It raises the following error:
```
error metro-config@0.77.0: The engine "node" is incompatible with this module. Expected version ">=18". Got "16.18.1"
```

### Test plan

As this doesn't add any feature, we need to make sure the core
functionality is not broken and all the CI checks pass.

---------

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>