fix: use `AndroidManifest.xml` for older RN versions
The React Native CLI relies on the filename `AndroidManifest.xml` to be able parse the package name.
But we were using `AndroidManifestLegacy.xml` instead which broke the lib for older RN versions (AGP 7).
This change makes it so that we use `AndroidManifest.xml` for older RN versions and `AndroidManifestNew.xml` for newer RN
versions.

Fixes #429