fix: use `AndroidManifest.xml` for older RN versions (#431)
### Summary

The React Native CLI relies on the filename `AndroidManifest.xml` to be
able parse the package name. But we were using
`AndroidManifestLegacy.xml` instead which broke the lib for older RN
versions (AGP 7). This change makes it so that we use
`AndroidManifest.xml` for older RN versions and `AndroidManifestNew.xml`
for newer RN versions.

Fixes #429

### Test plan

- Create an app with RN 0.68 and install the library without these
changes - see that `yarn android` fails
- Create an app with RN 0.68 and install the library with these changes
- see that `yarn android` succeeds
- Create an app with latest RN and install the library with these
changes - see that `yarn android` succeeds