refactor: bump the engine node version to 18 (#459)
### Summary

Node 16 LTS has [reached
EOL](https://nodejs.org/en/blog/announcements/nodejs16-eol). And some
libraries are dropping the support for it already. In order to support
the transition to Node 18 and make sure everything works properly, we
can change all the minimum engine versions to 18.

### Test plan

> This needs to be tested with a pure Javascript library which uses
Expo, and a library with native code which uses the CLI.

1. Create a new library using `npx create-react-native-library
some-cool-name`.
2. Complete creating the library and install the dependencies
3. Build the example app and make sure your library works properly.
4. Run `yarn prepack` to make sure your library is building properly.