fix: add abiFilters to the native template (#546)
<!-- Please provide enough information so that others can review your
pull request. -->
<!-- Keep pull requests small and focused on a single change. -->

### Summary

In the native template `abiFilters` is missing from
`externalNativeBuild` block. This means that the library would be built
for every available ABI when building the application, even if the
application itself is being built for a specific one. This is not that
problematic (besides taking more time) when using React Native aar, but
when building RN from source the core `.so` files will be built for that
specific ABI. In case the library depends on a `.so` file, it would fail
at the linking stage due to missing `.so` for the remaining ABIs.

### Test plan
See
- https://github.com/Expensify/App/pull/40102
- https://github.com/Expensify/react-native-live-markdown/pull/284
- https://github.com/margelo/react-native-quick-sqlite/pull/41

---------

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>