feat: support mts, cts, mjs and cts files in source code (#595)
### Summary

currently these files get treated as assets and are copied over as is.
however this is not the correct behavior. this changes them to be trated
as source code.

in addition, the module syntax in .mts or .mjs files is preserved as is.

### Test plan

tested in a generated project by creating `.mts` and `.cts` files.