feat: export babel and metro configs to reduce boilerplate (#600)
### Summary

Currently the example app has metro and babel configs setup to work
correctly in the library setup. However, the files are long and verbose
- increasing the maintenance cost as the consumer needs to manually
update them when we make any changes.

So this moves the configs to `react-native-builder-bob` package, so
users can import and use them. They still have the full flexibility to
customize anything they need.

### Test plan

Tested by creating an Expo app and app with native modules.