feat: add esm build option for typescript (#603)
This wraps up the ESM support to add support for TypeScript as well. For
TypeScript to work correctly, it's necessary to have 2 builds (for ESM
and CommonJS). With this, we generate one build and then copy the
folder, then add a `package.json` with `type` property for the module
system.