fix: single answer questions dont get stored with metadata (#633)
### Summary

If a question has a single answer on `create-react-native-library`, we
skip that question. However, due to a bug, single answer questions
weren't making it to the metadata stored with new project. This PR fixes
that

### Test plan

A good example is the `Fabric view with backward compat`. This option
had a single language setting `kotlin and objective c`. To test this PR,
you can create a new project with these options and make sure the
language is getting stored properly.