fix: spawn CLI directly without using Node (#689)
<!-- Please provide enough information so that others can review your
pull request. -->
<!-- Keep pull requests small and focused on a single change. -->

### Summary

![image](https://github.com/user-attachments/assets/2880ad03-3d96-4056-bbdf-299c34aba724)
The default is set to `yarn`, but an error occurs when using `pnpm` to
run `bob build`.

The error arises because the CLI is spawned via Node. Since the CLI in
`.bin` is designed to run with Node by default, it should spawn
correctly even without explicitly prepending node.
<!-- What existing problem does the pull request solve? Can you solve
the issue with a different approach? -->

### Test plan

> The changes work correctly even when tested with Yarn.

1. Create a repository using the command:
```sh
> npx create-react-native-library@latest rn-lib
```
3. Remove the `packageManager` field from the `package.json`.
4. Run `pnpm install`
* as-is
5. An error occurs.
* to-be
5. Success.

<!-- List the steps with which we can test this change. Provide
screenshots if this changes anything visual. -->

---------

Co-authored-by: Burak Güner <burak.guner@callstack.com>