refactor: drop backward compatible templates
As of React Native 0.76, the new architecture is now default.
So this moves the templates with new architecture up.
New libraries should start with new arch template.

For legacy usage, we still have the legacy template for now.
Old architecture libraries still work in new architecture via compat layer.
So the backward compatible templates are unnecessary.
They are also difficult to maintain and are currently broken, so I'm removing them.