feat: pass options to babel-preset using callers (#704)
This removes options parameter from the babel preset and passed them
using callers. With this change, Bob can automatically pass relevant
parameters to the preset when building without the user to do anything
additional if the chose to override the config.

**BREAKING CHANGE**

After this change, consumers can't provide options to the preset
directly.