fix: js-only libraries have codegen config (#758)
### Summary

Fixes #754

JS-only libraries were setting the new architecture flag on. While this
is not a preferable mental modal, we had some code associated with
codegen that checked this flag. With this PR, JS-only libraries will set
this flag to false. This is not an optimal solution but we will fully
drop this flag with the deprecation of the legacy arch in the upcoming
months.

### Test plan

1. Generate a JavaScript-only library
1. Make sure the `codegen` target isn't added to bob targets in
`package.json`
1. Make sure there is no `codegenConfig` in `package.json`
1. Make sure `yarn prepare` works.