refactor: drop cosmiconfig (#802)
### Summary

[cosmiconfig](https://github.com/cosmiconfig/cosmiconfig) is no longer
maintained. Apart from that, we can't support ESM configuration properly
because `cosmiconfig` only supports loading it asynchronously - but we
need to be able to load it synchronously in babel config.

I also took a look at
[lilconfig](https://github.com/antonk52/lilconfig), but it also has the
same limitation.

Node 20.19.0 onwards supports synchronous `require` of ESM, so this
limitation is not a technical limitation. So I decided to implement
config loading myself since we don't need any advanced features of these
config loaders such as traversing up until they find the config.

### Test plan

Tested in a sample project with both `bob build` and in the example app
with metro.