refactor: drop old arch and cpp templates (#818)
## Summarry

The large number of templates makes it difficult to maintain and test.
so this reduces the number of templates:

- Drop old architecture templates - users can use codegen or older
version of crnl
- Drop cpp templates - our cpp template isn't a pure cpp template, we
should eventually have it

## Test plan

Tested building each template: turbo module, fabric view, nitro module
and ran the example app for Android & iOS to verify it works.

- Nitro fails locally on iOS due to this
https://github.com/mrousavy/nitro/issues/422, so had to force the iOS
version to make it work
- Turbo module works fine locally on iOS but fails on CI

Both of them seem related to XCode 16 and not this change.