fix: Add `nitro.json` to npm package (#860)
In a future version of Nitro, you will be able to import types
(HybridObjects) from other libraries. For the import to work seamlessly,
nitrogen needs to read the other library's `nitro.json` so it knows
which C++/Android/Swift namespace to import to resolve the foreign type.
That's why the `nitro.json` file should be part of the npm package! :)