fix: fix `…Package.kt` being in wrong Java `package` for Nitro (#869)
<!-- Please provide enough information so that others can review your
pull request. -->
<!-- Keep pull requests small and focused on a single change. -->

### Summary

Previously, CRNL moved the `...Package.kt` into the wrong Java
`package`. The Java `package` from the `build.gradle` is expected.
This PR fixes that.

<!-- What existing problem does the pull request solve? Can you solve
the issue with a different approach? -->

### Test plan

Build a newly created Nitro library with CRNL. Do we have CI for that?
We should definitely add one
<!-- List the steps with which we can test this change. Provide
screenshots if this changes anything visual. -->

Co-authored-by: Satyajit Sahoo <satyajit.happy@gmail.com>